refactor(auto_search): convert to subdir + extract scoring.rs (Phase 1 start)
First checkpoint of the Phase 1 auto_search.rs split (6557 lines → 6
files). This commit does two things:

  1. Convert src/services/auto_search.rs → src/services/auto_search/mod.rs
     (pure file move, no content change).
  2. Extract the three scoring functions into a sibling:
       pub(super) fn apply_cf_seadex_overlay
       pub(super) fn format_scoring_detail
       pub(super) fn rescore_for_auto_search
     plus their 4 unit tests (format_scoring_detail_matches_plan_docs_example,
     _empty_breakdown_drops_bracket_section, _negative_cf_has_sign_and_marks_drop,
     _surfaces_seadex_bonus) into src/services/auto_search/scoring.rs.

scoring.rs imports from super via:
    use super::{SearchTarget, normalize_title, parse_release_numbers,
                season_mismatch, token_overlap_ratio, token_set};
The only visibility change in mod.rs: `season_mismatch` upgraded from
private `fn` to `pub(super) fn` so rescore_for_auto_search can reach
it from the sibling. Everything else in the closure was already pub.

No behaviour change. Tests pass (601 + 4 scoring-module tests = 605,
cargo reports 601 because the 4 moved tests were already in the count;
ran them separately via `cargo test format_scoring` to confirm they now
live at services::auto_search::scoring::tests).

mod.rs: 6557 → 6305 lines. scoring.rs: 287 lines.
Remaining siblings to extract in follow-up commits: search_target,
aliases, release_parse, pack_detection.
